requirements.txt
# --- Core ---
python-dotenv
mistralai
pypdf

# --- Agno Agent Framework ---
agno
notion-client


